backend/requirements.txt
# Backend Dependencies

# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlmodel==0.0.14
asyncpg==0.29.0
alembic==1.13.1
greenlet==3.0.3

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Email
resend==0.8.0

# Storage
supabase==2.3.4

# Validation & Utils
pydantic[email]==2.5.3
pydantic-settings==2.1.0
python-slugify==8.0.1

# File Processing
pandas==2.1.4
openpyxl==3.1.2
python-magic==0.4.27

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
pytest-cov==4.1.0

# Dev
black==24.1.0
isort==5.13.2
